+++ modules/dds/register_types.cpp
@@ -30,11 +30,9 @@
 
 #include "register_types.h"
 
-#include "image_loader_dds.h"
 #include "texture_loader_dds.h"
 
 static Ref<ResourceFormatDDS> resource_loader_dds;
-static Ref<ImageLoaderDDS> image_loader_dds;
 
 void initialize_dds_module(ModuleInitializationLevel p_level) {
 	if (p_level != MODULE_INITIALIZATION_LEVEL_SCENE) {
@@ -43,9 +41,6 @@ void initialize_dds_module(ModuleInitializationLevel p_level) {
 
 	resource_loader_dds.instantiate();
 	ResourceLoader::add_resource_format_loader(resource_loader_dds);
-
-	image_loader_dds.instantiate();
-	ImageLoader::add_image_format_loader(image_loader_dds);
 }
 
 void uninitialize_dds_module(ModuleInitializationLevel p_level) {
@@ -55,7 +50,4 @@ void uninitialize_dds_module(ModuleInitializationLevel p_level) {
 
 	ResourceLoader::remove_resource_format_loader(resource_loader_dds);
 	resource_loader_dds.unref();
-
-	ImageLoader::remove_image_format_loader(image_loader_dds);
-	image_loader_dds.unref();
 }
